Panache entity done, changes the id type to long

+++ src/main/java/anand/learn/UserResource.java
@@ -22,7 +22,7 @@ public class UserResource {
     @GET
     @Path("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response getUserById(@PathParam("id") int userId) {
+    public Response getUserById(@PathParam("id") long userId) {
         return User.findByIdOptional(userId)
                 .map(Response::ok)
                 .orElse(Response.status(Response.Status.NOT_FOUND)).build();
@@ -42,7 +42,7 @@ public class UserResource {
     @Path("/{id}")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
-    public Response updateUser(@PathParam("id") int existingUserId,
+    public Response updateUser(@PathParam("id") long existingUserId,
                                User newUser) {
         Optional<User> existingUserOptional = User.findByIdOptional(existingUserId);
         if(existingUserOptional.isPresent()) {
@@ -64,7 +64,7 @@ public class UserResource {
     @Transactional
     @Path("/{deleteUserId}")
     @Produces(MediaType.TEXT_PLAIN)
-    public Response deleteUser(@PathParam("deleteUserId") int deleteUserId) {
+    public Response deleteUser(@PathParam("deleteUserId") long deleteUserId) {
         boolean isDeleted = User.deleteById(deleteUserId);
         if (isDeleted) {
             return Response.ok("User deleted successfully").build();
